requirement.txt
python==3.9.19
numpy==1.26.4
python-dotenv==1.0.1
langchain==0.2.5
langchain-core==0.2.9
langchain-openai==0.1.8
langchain-anthropic==0.1.15
langchain-mistralai==0.1.8
fastapi==0.111.0
python-multipart==0.0.9
pytesseract==0.3.10
// also install the tesseract engine // 
// poppler # see pdf2image repository readme //
pdf2image==1.17.0
scikit-learn==1.5.0